src/main/java/org/sam/algo/containsduplicate/ContainsDuplicateSolution.java
package org.sam.algo.containsduplicate;

import java.util.Arrays;
import java.util.HashSet;

public class ContainsDuplicateSolution {
    public boolean containsDuplicate1(int[] nums) {
        Arrays.sort(nums);
        for (int i = 0; i < nums.length - 1; i++) {
            if (nums[i] == nums[i + 1]) return true;
        }
        return false;
    }

    public boolean containsDuplicate2(int[] nums) {
        HashSet<Integer> set = new HashSet<>();
        for (int i = 0; i < nums.length; i++) {

            if (set.contains(nums[i])) {
                return true;
            } else {
                set.add(nums[i]);
            }
        }
        return false;
    }
}
